refactor: banner 위치 변경
