+++ src/app/page.tsx
@@ -1,4 +1,4 @@
-import Banner from "./banner";
+import Banner from "./components/banner";
 
 export default function Home() {
   return (
